refactor to support streams and restructure files
