+++ crobat/Cargo.toml
@@ -11,7 +11,7 @@ prost = "0.6"
 webpki = "0.21.3"
 webpki-roots = "0.20.0"
 rustls = "0.18.0"
-futures-core = "0.3"
+futures-core = "0.3.8"
 futures-util = "0.3"
 async-stream = "0.2"
 log = "0.4"
